requirements.txt
# Core dependencies
pydantic-ai
pydantic-settings
aiohttp

# Testing
pytest
pytest-watch
pytest-asyncio
pytest-cov
pytest-vcr

# Linting & Formatting
black
isort
autoflake

# Optional: Development Tools
ipython
